+++ django_oemof/__init__.py
@@ -1,3 +1,3 @@
 """Init of django_oemof app, holds version"""
 
-VERSION = "0.11.0"
+VERSION = "0.12.0"